% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{appointments_billings_sameday}
\alias{appointments_billings_sameday}
\title{Billings done on same day as appointments or visit}
\usage{
appointments_billings_sameday(
  dMeasureBillings_obj,
  date_from,
  date_to,
  clinicians,
  lazy,
  screentag,
  screentag_print
)
}
\arguments{
\item{dMeasureBillings_obj}{dMeasureBillings R6 object}

\item{date_from}{(default $date_a from dMeasure) start date}

\item{date_to}{(default $date_b from dMeasure) end date (inclusive)}

\item{clinicians}{(default $clinicians from dMeasure) list of clinicians to view}

\item{lazy}{if TRUE, then do not recalculate appointment list. otherwise, re-calculate}

\item{screentag}{(default FALSE) optionally add a fomantic/semantic HTML description of 'action'}

\item{screentag_print}{(default TRUE) optionally add a 'printable' description of 'action'}
}
\value{
list of appointments (with patient details)
}
\description{
filter to billings which are done on the same day as displayed appointments
}
