% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{billed_appointments}
\alias{billed_appointments}
\title{List of appointments with billings}
\usage{
billed_appointments(
  dMeasureBillings_obj,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  status = NA,
  lazy = FALSE
)
}
\arguments{
\item{dMeasureBillings_obj}{dMeasureBillings R6 object}

\item{date_from}{start date, inclusive (date object)}

\item{date_to}{end date, inclusive (date object)
default of date_from and date_to defined by choose_date method of dMeasure}

\item{clinicians}{(default $clinicians of dMeasure object) list of clinicians to view}

\item{status}{(default NA) filter by 'status' if not NA
permissible values are 'Booked', 'Completed', 'At billing',
'Waiting', 'With doctor'}

\item{lazy}{(default FALSE) if lazy=TRUE, then don't re-calculate appointments_filtered to calculate}
}
\value{
list of appointments
}
\description{
Filtered by date, and chosen clinicians
}
\details{
Billings for patients who have displayed appointments

collects ALL billings for patients who have displayed appointments
used by CDM billings view
}
