% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userInterface.R
\name{datatableServer}
\alias{datatableServer}
\title{Billings module - server for UI}
\usage{
datatableServer(input, output, session, dMBillings)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{dMBillings}{dMeasureCustom R6 object}
}
\value{
none
}
\description{
Billings module - server for UI
}
