% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Billings.R
\docType{data}
\name{dMeasureBillings}
\alias{dMeasureBillings}
\title{dMeasureBillings class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
dMeasureBillings
}
\description{
list services (billings) by clinician providers
}
\details{
dMeasureBillings class
}
\keyword{datasets}
