% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{gpmp_list}
\alias{gpmp_list}
\title{List whether patients had a recent GPMP (care plan)}
\usage{
gpmp_list(
  dMeasureBillings_obj,
  appointments = NULL,
  months_min = 0,
  months_max = 12
)
}
\arguments{
\item{dMeasureBillings_obj}{dMeasureBillings R6 object}

\item{appointments}{dataframe of appointments $InternalID and $Date
if no parameter provided, derives from $appointments_filtered}

\item{months_min}{minimum age in months (inclusive, 'from')}

\item{months_max}{maximum age in months (exclusive, 'up to')}
}
\value{
a vector of numbers, whih care internalID
}
\description{
List whether patients had a recent GPMP (care plan)
}
