% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{list_services_allclinicians}
\alias{list_services_allclinicians}
\title{List of services with date of birth}
\usage{
list_services_allclinicians(
  dMeasureBillings_obj,
  date_from = NA,
  date_to = NA,
  internalID = NA,
  payerCode = NA
)
}
\arguments{
\item{dMeasureBillings_obj}{dMeasureBillings R6 object}

\item{date_from}{default is $date_a. start date, inclusive (date object)}

\item{date_to}{default is $date_b end date, inclusive (date object)}

\item{internalID}{default is NA (include all). otherwise, a vector of internal IDs to include (filter)}

\item{payerCode}{default is $payerCode
0 = unknown
1 = private (patient)
2 = Medicare direct billing 'bulk-billing'
3 = Department of Veteran Affairs 'DVA'
4 = WorkCover
5 = private (head of family)
8 = private (other)}
}
\value{
list of services
}
\description{
Filtered by date. Not filtered by clinicians
}
