% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Billings.R
\name{own_billings}
\alias{own_billings}
\title{show only 'own' billings next to provider name}
\usage{
own_billings(dMeasureBillings_obj, value)
}
\arguments{
\item{dMeasureBillings_obj}{dMeasureBillings R6 object}
}
\value{
current value of own_billings if no parameter provided
}
\description{
show only 'own' billings next to provider name
}
