% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{tag_billings_list}
\alias{tag_billings_list}
\title{tags a billlings list for printing or HTML screen display}
\usage{
tag_billings_list(billings_list, screentag = FALSE, screentag_print = TRUE)
}
\arguments{
\item{billings_list}{dataframe. requires ServiceDate and
lists (MBSItem, MBSDescription, MBSProvider)}

\item{screentag}{(default FALSE) optionally add a fomantic/semantic HTML description of 'action'}

\item{screentag_print}{(default TRUE) optionally add a 'printable' description of 'action'}

\item{dMeasureBillings_obj}{dMeasureBillings R6 object}
}
\value{
tagged billings list
 adds billingtag_print for 'print' view
 adds billingtag for HTML view using fomantic/semantic buttons
}
\description{
tags a billlings list for printing or HTML screen display
}
