% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userInterface.R
\name{datatableServer}
\alias{datatableServer}
\title{Billings module - server for UI}
\usage{
datatableServer(id, dMBillings)
}
\arguments{
\item{id}{id string}

\item{dMBillings}{dMeasureCustom R6 object}
}
\value{
none
}
\description{
Billings module - server for UI
}
