% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Billings.R
\name{dMeasureBillings}
\alias{dMeasureBillings}
\title{dMeasureBillings class}
\description{
list services (billings) by clinician providers
}
\details{
dMeasureBillings class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{dMeasureBillings$new()}}
\item \href{#method-billed_appointments}{\code{dMeasureBillings$billed_appointments()}}
\item \href{#method-list_services}{\code{dMeasureBillings$list_services()}}
\item \href{#method-list_services_allclinicians}{\code{dMeasureBillings$list_services_allclinicians()}}
\item \href{#method-appointments_billings_sameday}{\code{dMeasureBillings$appointments_billings_sameday()}}
\item \href{#method-list_billings}{\code{dMeasureBillings$list_billings()}}
\item \href{#method-clone}{\code{dMeasureBillings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureBillings$new(dMeasure_obj)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-billed_appointments"></a>}}
\subsection{Method \code{billed_appointments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureBillings$billed_appointments(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  status = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_services"></a>}}
\subsection{Method \code{list_services()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureBillings$list_services(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  internalID = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_services_allclinicians"></a>}}
\subsection{Method \code{list_services_allclinicians()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureBillings$list_services_allclinicians(
  date_from = NA,
  date_to = NA,
  internalID = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-appointments_billings_sameday"></a>}}
\subsection{Method \code{appointments_billings_sameday()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureBillings$appointments_billings_sameday(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  lazy = FALSE,
  screentag = FALSE,
  screentag_print = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_billings"></a>}}
\subsection{Method \code{list_billings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureBillings$list_billings(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  own_billings = NA,
  lazy = FALSE,
  screentag = FALSE,
  screentag_print = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureBillings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
