% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{list_billings}
\alias{list_billings}
\title{Billings, appointments and visit combined view}
\usage{
list_billings(
  dMeasureBillings_obj,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  own_billings = NA,
  lazy = FALSE,
  payerCode = NA
)
}
\arguments{
\item{dMeasureBillings_obj}{dMeasureBillings R6 object}

\item{date_from}{(default date_a field) start date}

\item{date_to}{(default date_b field) end date (inclusive)}

\item{clinicians}{(default clinicians field) list of clinicians to view}

\item{own_billings}{(default own_billings field) logical TRUE/FALSE

 only show billings done by the specified provider on the
 contact (visit/appointment/billing) day.

 if FALSE, the all billings
 done on the patient on that day are shown.}

\item{lazy}{if TRUE, then do not recalculate appointment list. otherwise, re-calculate}

\item{payerCode}{default is $payerCode
0 = unknown
1 = private (patient)
2 = Medicare direct billing 'bulk-billing'
3 = Department of Veteran Affairs 'DVA'
4 = WorkCover
5 = private (head of family)
8 = private (other)}

\item{screentag}{(default FALSE) optionally add a fomantic/semantic HTML description of 'action'}

\item{screentag_print}{(default TRUE) optionally add a 'printable' description of 'action'}
}
\value{
list of appointments (with patient details) and attached billings
 billings are in list() of MBSItem, Description and Provider
}
\description{
filter to chosen date range and practitioners
}
\details{
billings are aggregated/group to patient/day
}
