% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{list_services_allclinicians}
\alias{list_services_allclinicians}
\title{List of services with date of birth}
\usage{
list_services_allclinicians(dMeasureBillings_obj, date_from = NA,
  date_to = NA, lazy = FALSE)
}
\arguments{
\item{dMeasureBillings_obj}{dMeasureBillings R6 object}

\item{date_from}{default is $date_a. start date, inclusive (date object)}

\item{date_to}{default is $date_b end date, inclusive (date object)}

\item{lazy}{if lazy=TRUE, then don't re-calculate $appointments_filtered to calculate}
}
\value{
list of services
}
\description{
Filtered by date. Not filted by clinicians
}
